pyproject.toml
[tool.poetry]
name = "prompts"
version = "0.1.0"
description = ""
authors = ["dsthiago <thiagosd.2.18@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
# anthropic = "^0.39.0"
# ipykernel = "^6.29.5"
# ipywidgets = "^8.1.5"
# python-dotenv = "^1.0.1"
# openai = "^1.54.5"
# crewai = {extras = ["tools"], version = "^0.95.0"}
ipykernel = "^6.29.5"
ipywidgets = "^8.1.5"
python-dotenv = "^1.0.1"
openai = "^1.58.1"
anthropic = "^0.42.0"
langchain-anthropic = "^0.3.1"
duckduckgo-search = "^6.4.2"
langchain-openai = "^0.2.14"
opentelemetry-api = "^1.29.0"
opentelemetry-sdk = "^1.29.0"
langchain = "^0.3.13"
numpy = "<2"
langchain-core = "^0.3.28"
pydantic = "^2.10.4"
langchain-community = "^0.3.13"
unstructured = "^0.16.11"
pypandoc = "^1.14"
voyageai = "^0.3.2"
selenium = "^4.27.1"
beautifulsoup4 = "^4.12.3"
webdriver-manager = "^4.0.2"
bs4 = "^0.0.2"
geckodriver-autoinstaller = "^0.1.0"
maritalk = "^0.2.6"
torch = ">=2.0.0, !=2.0.1, !=2.1.0" # https://stackoverflow.com/a/76647180
lightning = "^2.1.0"
pytorchtools = "^0.0.2"
peft = "^0.14.0"
datasets = "^3.2.0"
transformers = "^4.48.2"
accelerate = "^1.3.0"
python-docx = "^1.1.2"
rouge-score = "^0.1.2"
bitsandbytes-cuda110 = "^0.26.0.post2"
bitsandbytes = "^0.45.1"
transformers-stream-generator = "^0.0.5"
einops = "^0.8.0"
trl = "^0.14.0"
flash-attn = "^2.7.4.post1"
wandb = "^0.19.5"
absl-py = "^2.1.0"
nltk = "^3.9.1"
pypdf2 = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
